% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare_cols}
\alias{compare_cols}
\title{Compare column names of two data frames to find differences}
\usage{
compare_cols(c1, c2)
}
\arguments{
\item{c1, c2}{data frames to compare}
}
\value{
a named list with 4 elements:
\itemize{
\item \code{same}: a logical indicating whether \code{c1} and \code{c2} have the same columns
\item \code{both}: a character vector with the column names in both \code{c1} and \code{c2}
\item \code{c1_only}: a character vector with the column names only in \code{c1}
\item \code{c2_only}: a character vector with the column names only in \code{c2}
}
}
\description{
Indicates if two data frames have the same column names or not and what the
differences are.
}
