% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compareDF_custom}
\alias{compareDF_custom}
\title{Compare two data frames using \code{\link[compareDF:compare_df]{compareDF::compare_df()}}}
\usage{
compareDF_custom(c1, c2, id_cols, tolerance = 1e-05)
}
\arguments{
\item{c1, c2}{data frames to compare}

\item{id_cols}{a character vector of id columns output from \code{\link[=get_uid_cols]{get_uid_cols()}}}

\item{tolerance}{amount by which numeric values can vary and still be
considered equal (helps reduce machine rounding errors false positive
differences)}
}
\value{
a list returned by \code{\link[compareDF:compare_df]{compareDF::compare_df()}}
}
\description{
A convenience wrapper for \code{\link[compareDF:compare_df]{compareDF::compare_df()}}
}
\details{
Rows that were duplicated in \code{c1} or \code{c2} should be removed prior to calling
this function
}
