% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareDFx-package.R
\docType{package}
\name{compareDFx-package}
\alias{compareDFx}
\alias{compareDFx-package}
\title{compareDFx: Compare Data Frame Versions, a CompareDF Extension}
\description{
Compares two data frames with a detailed Excel report
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/skgithub14/compareDFx}
  \item Report bugs at \url{https://github.com/skgithub14/compareDFx/issues}
}

}
\author{
\strong{Maintainer}: Stephen Knapp \email{stephen@knappconsultingllc.com} (\href{https://orcid.org/0000-0002-5101-4555}{ORCID})

}
\keyword{internal}
