% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{create_comparison_excel}
\alias{create_comparison_excel}
\title{Create Excel report for comparing two data frames}
\usage{
create_comparison_excel(comparison, path = "comparison_report.xlsx")
}
\arguments{
\item{comparison}{a list returned from \code{\link[=get_comparison]{get_comparison()}}}

\item{path}{a string, path of the Excel file to be written, default is
\code{"comparison_report.xlsx"} (must have .xlsx suffix)}
}
\value{
nothing
}
\description{
Creates a multi-tab report for comparing two data frames with one
tab for each element of an object returned by \code{\link[=get_comparison]{get_comparison()}}
}
