% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df2}
\alias{df2}
\title{Example Data Frame 2}
\format{
\subsection{'df2'}{

A data frame with 8 rows and 12 columns:
\describe{
\item{id1}{ID column 1}
\item{id2}{ID column 2}
\item{num}{a numeric column}
\item{char}{a character column}
\item{int}{an integer column}
\item{log}{a logical column}
\item{date}{a Date column}
\item{class_num_char}{a character column that corresponds to a numeric column in \code{df1}}
\item{class_num_log}{a logical column that corresponds to a numeric column in \code{df1}}
\item{dec_diff_ref}{a numeric column used to calculate \code{dec_diff}}
\item{dec_diff}{a numeric column for testing \code{tolerance} sensitivity}
\item{extra2}{a column in \code{df2} which does not exist in \code{df1}}
}
}
}
\usage{
df2
}
\description{
An example data frame for testing compareDFx
}
\keyword{datasets}
