% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{find_dups}
\alias{find_dups}
\title{Separate duplicated data frame rows from rows with no duplicates}
\usage{
find_dups(df, id_cols)
}
\arguments{
\item{df}{a data frame with or without duplicated rows}

\item{id_cols}{a character vector of the column names that form a unique ID}
}
\value{
a names list with four elements:
\itemize{
\item \code{every_dups}: rows which were duplicated for every value, there will be
one distinct row in \code{every_dups} for each case and the column \code{n} will
list how many occurences of the row were in \code{df}.
\item \code{id_dups}: rows which were duplicated based on \code{id_cols} but had other
values that differed, the column \code{n} contains how many ID duplicates of
one ID exist in \code{df}
\item \code{id_NA}: rows which had an \code{NA} value in a \code{id_cols} column
\item \code{no_dups}: this is \code{df} with all \code{id_dups} row removed and only one copy
of each \code{every_dups} row.
}
}
\description{
Separates a data frame into sub-data frames based on types of row duplication
or lack thereof. Also identifies \code{NA} values in \code{id_cols}.
}
