% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_comparison.R
\name{get_changed_row_indices_by_column}
\alias{get_changed_row_indices_by_column}
\title{Get row indices by column of changed values}
\usage{
get_changed_row_indices_by_column(
  top_bottom_view,
  left_right_view,
  comparison_table_diff,
  id_cols
)
}
\arguments{
\item{top_bottom_view}{a data frame of comparison data where each row in
\code{df1} is displayed directly above its \code{df2} counterpart}

\item{left_right_view}{a data frame of comparison data where each column in
\code{df1} is displayed directly to the left of its \code{df2} counterpart}

\item{comparison_table_diff}{a data frame returned by the
\code{comparison_table_diff} element of a \code{\link[=compare_df_wrapper]{compare_df_wrapper()}} output}

\item{id_cols}{a character vector of id columns output from \code{\link[=compare_cols]{compare_cols()}}}
}
\value{
a named list where each element's name is a column in
\code{comparison_data} and the elements themselves are numeric vectors with the
row indices that changed
}
\description{
Given both top-bottom and left-right comparison data frames, this function
determines the row indices for each comparison data frame, by column, that
have changed between \code{df1} and \code{df2}
}
