% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_comparison.R
\name{get_left_right_view}
\alias{get_left_right_view}
\title{Convert data comparison from top-bottom to left-right view}
\usage{
get_left_right_view(top_bottom_view, id_cols)
}
\arguments{
\item{top_bottom_view}{a data frame of comparison data where each row in
\code{df1} is displayed directly above its \code{df2} counterpart}

\item{id_cols}{a character vector of id columns output from \code{\link[=compare_cols]{compare_cols()}}}
}
\value{
a modified copy of \code{top_bottom_view} that displays \code{df1} data to
the left of the \code{df2} data instead of above the \code{df2} data
}
\description{
Takes a comparison data frame produced in \code{\link[=get_comparison]{get_comparison()}} and manipulates
it so that it displays each \code{df1} columns to the left of its \code{df2}
counterpart.
}
