% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_comparison.R
\name{get_pivoted_view}
\alias{get_pivoted_view}
\title{Get pivot comparison view}
\usage{
get_pivoted_view(top_bottom_view, top_bottom_view_change_indices, id_cols)
}
\arguments{
\item{top_bottom_view}{a data frame of comparison data where each row in
\code{df1} is displayed directly above its \code{df2} counterpart}

\item{top_bottom_view_change_indices}{the \code{top_bottom} output of the
\code{\link[=get_changed_row_indices_by_column]{get_changed_row_indices_by_column()}} which provides the column and rows in
\code{top_bottom_view} that have changed between \code{df1} and \code{df2}}

\item{id_cols}{a character vector of id columns output from \code{\link[=compare_cols]{compare_cols()}}}
}
\value{
a data frame
}
\description{
Creates a comparison view of the data which lists out every value in \code{df1}
directly next to its corresponding value in \code{df2} in its own row. Each row in
this view is uniquely identified by id columns in \code{df1} and \code{df2} along with
a column of column names from \code{df1} and \code{df2}. This view also includes a
\verb{record from} column which described whether the records in the row was from
\code{df1}, \code{df2}, or \verb{df1 and df2}. It also includes a \verb{Value Changed} column
which will be marked as \code{"Y"} in the discrepancy change for a record that
was an addition, deletion, or change, marked as \code{"N"} for matched records,
or marked as \code{"UNK"} for records that could not be compared between \code{df1} and
\code{df1}.
}
