% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_comparison.R
\name{get_top_bottom_view}
\alias{get_top_bottom_view}
\title{Segregates data frame comparisons by type}
\usage{
get_top_bottom_view(
  comparison_df,
  comparison_table_diff,
  id_cols,
  cc_out,
  df1_exact_dups,
  df2_exact_dups,
  df1_id_dups,
  df2_id_dups,
  df1_id_NA,
  df2_id_NA
)
}
\arguments{
\item{comparison_df}{the \code{comparison_df} element of a \code{\link[=compare_df_wrapper]{compare_df_wrapper()}}
output}

\item{comparison_table_diff}{the \code{comparison_table_diff} element of a
\code{\link[=compare_df_wrapper]{compare_df_wrapper()}} output}

\item{id_cols}{a character vector of id columns output from \code{\link[=compare_cols]{compare_cols()}}}

\item{cc_out}{a list output from \code{\link[=compare_cols]{compare_cols()}}}

\item{df1_exact_dups, df2_exact_dups}{an \code{exact_dups} list element from a
\code{\link[=find_dups]{find_dups()}} output, for \code{df1} and \code{df2}, respectively}

\item{df1_id_dups, df2_id_dups}{an \code{id_dups} list element from a \code{\link[=find_dups]{find_dups()}}
output, for \code{df1} and \code{df2}, respectively}

\item{df1_id_NA, df2_id_NA}{an \code{id_NA} list element from a \code{\link[=find_dups]{find_dups()}}
output, for \code{df1} and \code{df2}, respectively}
}
\value{
a data frame with all data from \code{df1} and \code{df2} with comparison
annotations where each row in \code{df1} is displayed directly above its \code{df2}
counterpart
}
\description{
Splits the output of \code{\link[=compare_df_wrapper]{compare_df_wrapper()}} in several smaller data frames
organized by type of change and formats it
}
