% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{summarize_compare_cols}
\alias{summarize_compare_cols}
\title{Summary metrics for data frame comparison by columns}
\usage{
summarize_compare_cols(
  df1,
  df2,
  id_cols,
  cc_out,
  df_standard_cols,
  comparison_table_diff
)
}
\arguments{
\item{df1, df2}{data frames to compare}

\item{id_cols}{a character vector of id columns output from \code{\link[=get_uid_cols]{get_uid_cols()}}}

\item{cc_out}{a list output from \code{\link[=compare_col]{compare_col()}}}

\item{df_standard_cols}{a list returned from \code{\link[=standardize_col_order]{standardize_col_order()}}}

\item{comparison_table_diff}{the \code{comparison_table_diff} list element of a
\code{\link[=compare_df_wrapper]{compare_df_wrapper()}} output.}
}
\value{
a named list containing two data frames:
\itemize{
\item \code{report_simple} a data frame with higher level summary statistics
\item \code{report_by_col} a data frame that lists summary statistics by column name
}
}
\description{
Creates two data frames that summarize the differences by columns for two
versions of the same data frame. The first provides higher level summary
statistics on the number of columns in each data frame, the number of columns
with changes, and the number of columns that have the same name but
different classes. The second data frame lists by column the number of
changes, whether the column was included in the new and/or old data frame,
and the original column class.
}
