% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_comparison.R
\name{summarize_compare_rows}
\alias{summarize_compare_rows}
\title{Summary metrics for data frame comparison by rows}
\usage{
summarize_compare_rows(df1, df2, all, id_cols)
}
\arguments{
\item{df1, df2}{data frames to compare}

\item{all}{a data frame with all the compared data with \code{discrepancy} column}

\item{id_cols}{a character vector of id columns output from \code{\link[=compare_cols]{compare_cols()}}}
}
\value{
a data frame with three columns: \code{Rows}, \code{df1}, \code{df2}
}
\description{
Produces a table of high level summary statistics for comparing two data
frames in terms of rows
}
