% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_comparison.R
\name{compare_cols}
\alias{compare_cols}
\title{Compare column names of two data frames to find differences}
\usage{
compare_cols(df1, df2)
}
\arguments{
\item{df1, df2}{data frames to compare}
}
\value{
a named list with 4 elements:
\itemize{
\item \code{same}: a logical indicating whether \code{df1} and \code{df2} have the same columns
\item \code{both}: a character vector with the column names in both \code{df1} and \code{df2}
\item \code{df1_only}: a character vector with the column names only in \code{df1}
\item \code{df2_only}: a character vector with the column names only in \code{df2}
}
}
\description{
Indicates if two data frames have the same column names or not and what the
differences are.
}
