% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareDFx-package.R
\docType{package}
\name{compareDFx-package}
\alias{compareDFx}
\alias{compareDFx-package}
\title{compareDFx: Compare Data Frame Versions in R and MS Excel}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Compares two data frames with detailed reporting. The report can be analyzed natively in R or in MS Excel. The comparison analysis shows changes and errors by row and column and provides summary statistics. compareDFx can also compare data frames with different columns and column classes. The MS Excel report is a multi-tab workbook that color codes additions, deletions, and changes which can easily be used by a non-R users.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/skgithub14/compareDFx}
  \item \url{https://skgithub14.github.io/compareDFx/}
  \item Report bugs at \url{https://github.com/skgithub14/compareDFx/issues}
}

}
\author{
\strong{Maintainer}: Stephen Knapp \email{stephen@knappconsultingllc.com} (\href{https://orcid.org/0000-0002-5101-4555}{ORCID})

}
\keyword{internal}
