% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_comparison_excel.R
\name{create_comparison_excel}
\alias{create_comparison_excel}
\title{Create Excel report for comparing two data frames}
\usage{
create_comparison_excel(
  comparison,
  path = "comparison_report.xlsx",
  autoOpen = FALSE,
  df_names = NULL
)
}
\arguments{
\item{comparison}{a list returned from \code{\link[=get_comparison]{get_comparison()}}}

\item{path}{a string, path of the Excel file to be written, default is
\code{"comparison_report.xlsx"} (must have .xlsx suffix)}

\item{autoOpen}{a logical indicating whether the Excel file should be
automatically opened, default is \code{FALSE}}

\item{df_names}{an optional character vector of length two that provides
names for the data frames being compared, the first string in the vector is
for \code{df1} and the second string in the vector is for \code{df2}. If provided,
these names will appear in the Excel 'summary' tab. The default is \code{NULL}.}
}
\value{
nothing
}
\description{
Creates a multi-tab report for comparing two data frames with one tab for
each element of an object returned by \code{\link[=get_comparison]{get_comparison()}}
}
\examples{
id_cols <- c("id1", "id2")
comparison <- get_comparison(compareDFx::df1, compareDFx::df2, id_cols)
path <- tempfile(pattern = "comparison_report", fileext = ".xlsx")
create_comparison_excel(comparison, path)

}
