% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{get_comparison}
\alias{get_comparison}
\title{Compare two data frames}
\usage{
get_comparison(df1, df2, id_cols, tolerance = 1e-05)
}
\arguments{
\item{df1, df2}{data frames to compare}

\item{id_cols}{the column names in \code{df1} and \code{df2} that make up a unique ID}

\item{tolerance}{the amount by which two numbers can differ to be considered
equal, default is \code{0.00001}}
}
\value{
a named list of data frames with elements:
\itemize{
\item \code{col_summary_simple}: high level summary statistics of columns
\item \code{col_summary_by_col}: comparison summary statistics by column name
\item \code{row_summary}: comparison summary statistics by row
\item \code{all}: A data from the other list elements in one data frame
\item \code{matched}: matching rows
\item \code{adds}: added rows
\item \code{dels}: deleted rows
\item \code{changed}: changed rows, top-bottom view
\item \code{changed_lr}: changed rows, left-right view
\item \code{exact_dups}: exact duplicate rows
\item \code{id_dups}: ID duplicate rows
\item \code{id_NA}: Rows with with \code{NA} ID values
\item \code{df1}: The original df1
\item \code{df2}: The original df2
}
}
\description{
Generates a detailed comparison report of two data frames
}
