% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{insert_dummy_cols}
\alias{insert_dummy_cols}
\title{Force two data frames to have the same columns by creating dummy columns}
\usage{
insert_dummy_cols(df1, df2, cc_out, id_cols)
}
\arguments{
\item{df1, df2}{data frames to compare}

\item{cc_out}{a list output from \code{\link[=compare_cols]{compare_cols()}}}

\item{id_cols}{a character vector of id columns output from \code{\link[=compare_cols]{compare_cols()}}}
}
\value{
a named list with two elements:
\itemize{
\item \code{df1}: the modified \code{df1} data frame
\item \code{df2}: the modified \code{df2} data frame
}
}
\description{
For each column in data frame 1 that is missing in data frame 2, this
function creates the missing column in data frame 2 and fills it with \code{NA},
and vice-version. It also makes the classes dummy columns match and ensure
both data frames have the same column order.
}
