% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{segregate_compare}
\alias{segregate_compare}
\title{Segregates data frame comparisons by type}
\usage{
segregate_compare(comparison_df, id_cols)
}
\arguments{
\item{comparison_df}{the \code{comparison_df} element of a \code{\link[=compare_dfx]{compare_dfx()}}
output}

\item{id_cols}{a character vector of id columns output from \code{\link[=get_uid_cols]{get_uid_cols()}}}
}
\description{
Splits the output of \code{\link[=compare_dfx]{compare_dfx()}} in several smaller data frames
organized by type of change and formats it
}
