% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{segregate_compare}
\alias{segregate_compare}
\title{Segregates data frame comparisons by type}
\usage{
segregate_compare(
  comparison_df,
  comparison_table_diff,
  id_cols,
  cc_out,
  df1_exact_dups,
  df2_exact_dups,
  df1_id_dups,
  df2_id_dups,
  df1_id_NA,
  df2_id_NA,
  standard_col_list,
  df1,
  df2
)
}
\arguments{
\item{comparison_df}{the \code{comparison_df} element of a \code{\link[=compare_df_wrapper]{compare_df_wrapper()}}
output}

\item{comparison_table_diff}{the \code{comparison_table_diff} element of a
\code{\link[=compare_df_wrapper]{compare_df_wrapper()}} output}

\item{id_cols}{a character vector of id columns output from \code{\link[=compare_cols]{compare_cols()}}}

\item{cc_out}{a list output from \code{\link[=compare_cols]{compare_cols()}}}

\item{df1_exact_dups, df2_exact_dups}{an \code{exact_dups} list element from a
\code{\link[=find_dups]{find_dups()}} output, for \code{df1} and \code{df2}, respectively}

\item{df1_id_dups, df2_id_dups}{an \code{id_dups} list element from a \code{\link[=find_dups]{find_dups()}}
output, for \code{df1} and \code{df2}, respectively}

\item{df1_id_NA, df2_id_NA}{an \code{id_NA} list element from a \code{\link[=find_dups]{find_dups()}}
output, for \code{df1} and \code{df2}, respectively}

\item{standard_col_list}{a list returned from \code{\link[=standardize_col_order]{standardize_col_order()}}}

\item{df1, df2}{data frames to compare}
}
\value{
a named list with elements:
\itemize{
\item \code{col_summary_simple}: high level summary statistics of columns, data frame
\item \code{col_summary_by_col}: comparison summary statistics by column name, data frame
\item \code{row_summary}: comparison summary statistics by row, data frame
\item \code{all}: data frame with all data from \code{df1} and \code{df2}
\item \code{all_index_by_col}: a list of row indices by column of \code{all} where
\code{df1} and \code{df2} have different values
\item \code{matched}: data frame with rows of \code{df1} and \code{df2} that match exactly
\item \code{adds}: data frame with additions in \code{df1}
\item \code{dels}: data frame with deletions from \code{df2}
\item \code{changed}: data frame that compares \code{df1} and \code{df2} changed rows,
a top-bottom view
\item \code{change_index_by_col}: a list of row indices by column of \code{changed} where
\code{df1} and \code{df2} have different values
\item \code{changed_lr}: data frame that compares \code{df1} and \code{df2} changed rows,
a left-right view
\item \code{change_index_by_col_lr}: a list of row indices by column of \code{changed_lr}
where \code{df1} and \code{df2} have different values
\item \code{exact_dups}: a data frame that lists rows that are exact duplicates of
each other, by \code{df1} and \code{df2}
\item \code{id_dups}: a data frame that lists rows that have duplicated ID column
values (meaning IDs are not unique in \code{df1} or \code{df2}) but which are not
exact duplicates (at least one non-ID column is different from it duplicate
row)
\item \code{id_NA}: a data frame of rows with with \code{NA} ID values from \code{df1} and \code{df2}
\item \code{df1}: the original \code{df1} data frame
\item \code{df2}: the original \code{df2} data frame
\item \code{id_cols}: a vector of column names that collectively make a unique row ID
}
}
\description{
Splits the output of \code{\link[=compare_df_wrapper]{compare_df_wrapper()}} in several smaller data frames
organized by type of change and formats it
}
