% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{segregate_compare}
\alias{segregate_compare}
\title{Segregates data frame comparisons by type}
\usage{
segregate_compare(comp_comparison_df)
}
\arguments{
\item{comp_comparison_df}{the \code{comparison_df} element of a
\code{\link[=compareDF_custom]{compareDF_custom()}} output}
}
\description{
Splits the output of \code{\link[=compareDF_custom]{compareDF_custom()}} in several smaller data frames
organized by type of change and formats it
}
