% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{standardize_col_order}
\alias{standardize_col_order}
\title{Standardize column order}
\usage{
standardize_col_order(df1, df2, id_cols)
}
\arguments{
\item{df1, df2}{data frames to compare}

\item{id_cols}{a character vector of id columns output from \code{\link[=get_uid_cols]{get_uid_cols()}}}
}
\value{
a list with two elements:
\itemize{
\item \code{df1}: a modified data frame \code{df1}
\item \code{df2}: a modified data frame \code{df2}
}
}
\description{
Puts the columns of two data frames in the same order and places the unique
identifiers columns to the far left of the data frame. Data frames must have
the exact same columns (if not, apply \code{\link[=insert_dummy_cols]{insert_dummy_cols()}} first)
}
